refactor : event 및 repository 패키지 분리
