+++ src/main/java/kitchenpos/menu/event/VerifiedMenusEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.event;
+package kitchenpos.menu.event;
 
 import java.util.List;
