+++ src/main/java/kitchenpos/order/event/VerifiedAbleToChangeEmptyEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.event;
+package kitchenpos.order.event;
 
 public class VerifiedAbleToChangeEmptyEvent {
