+++ src/main/java/kitchenpos/order/event/VerifiedAbleToUngroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.event;
+package kitchenpos.order.event;
 
 import java.util.List;
