+++ src/main/java/kitchenpos/product/event/VerifiedMenuProductsEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.event;
+package kitchenpos.product.event;
 
 import java.math.BigDecimal;
 import java.util.List;
