+++ src/main/java/kitchenpos/table/repository/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.table.repository;
 
 import java.util.List;
 import java.util.Optional;
